requirements.txt
flask
numpy
opencv-python
pillow
pyproj
requests
Flask-Caching
waitress